feat(commands): attach /command as durable prefix chip instead of pasting (T654)
Clicking a /command now attaches it as a chip (like a file), stored
durably in the browser per thread (localStorage cp-cmd-<agent>-<thread>),
at most one at a time. On Send the command's prompt body is prepended as a
prefix to the message, then detached. The Edit + X buttons show only on the
attached chip; friendly non-technical tooltips explain the behaviour.

Frontend-only, mirrored across desktop + mobile twins. useComposer gains
commandKey + attachedCmd state + attachCmd/detachCmd/persistCmd. Tab/Space
now attach the matched command. ComposerInputRow + Banner extracted into
fileUpload/composerInput.tsx (both twins) to keep ThreadComposer < 500 lines.
