+++ web/src/mobile-components/threads/ThreadComposer.tsx
@@ -1,14 +1,17 @@
 import { useEffect, useMemo, useRef, useState } from "react"
-import { animate, createSpring } from "animejs"
-import { lineBounds, resolveEnter, resolveTab, prefersReducedMotion } from "@/lib/utils"
+import { lineBounds, resolveEnter, resolveTab } from "@/lib/utils"
 import { measure } from "@/lib/support/telemetry"
-import { ArrowUp, Plus, Loader2, Clock, Pause } from "lucide-react"
 import type { ThreadStatus } from "@/lib/types"
 import { ComposerBubbles } from "@/mobile-components/threads/fileUpload"
 import type {
   UploadedFile,
   CommandSuggestion,
 } from "@/mobile-components/threads/fileUpload/helpers"
+import {
+  ComposerInputRow,
+  ComposerBanner,
+  type Banner,
+} from "@/mobile-components/threads/fileUpload/composerInput"
 import { FrostedBottomBar } from "@/mobile-components/shell/chrome/FrostedBottomBar"
 import { parseDraft } from "@/lib/support/threadMessages"
 
@@ -16,14 +19,6 @@ import { parseDraft } from "@/lib/support/threadMessages"
 // composer pill families share ONE module + ONE rendered row). Re-exported for
 // the mobile ThreadConversation consumer, matching the desktop twin's surface.
 export type { CommandSuggestion } from "@/mobile-components/threads/fileUpload/helpers"
-
-/** The turn-status banner shown above the composer input, or null. */
-interface Banner {
-  working: boolean
-  paused: boolean
-  color: string | undefined
-  text: string
-}
 
 /**
  * Resolve the composer's turn-status banner from thread state (T39/T371). Flat
@@ -75,12 +70,36 @@ interface Composer {
   caret: number
   textareaRef: React.RefObject<HTMLTextAreaElement | null>
   slashPrefix: string | null
+  /** the currently-attached `/command` (prepended to the message on send), or
+   *  null. Durable per-thread (T654). */
+  attachedCmd: CommandSuggestion | null
   canSend: (pendingFiles: number) => boolean
   onChange: (e: React.ChangeEvent<HTMLTextAreaElement>) => void
   onSelect: (e: React.SyntheticEvent<HTMLTextAreaElement>) => void
   handleKeyDown: (e: React.KeyboardEvent<HTMLTextAreaElement>) => void
   handleSubmit: () => void
-  prefill: (s: CommandSuggestion) => void
+  /** Attach a `/command` as the message prefix (replaces any current one, max
+   *  one) and strip the slash trigger line from the draft. */
+  attachCmd: (s: CommandSuggestion) => void
+  /** Remove the attached `/command`. */
+  detachCmd: () => void
+}
+
+/** Read the persisted attached command for a thread (T654). Tolerant: a missing
+ *  or malformed value yields null. */
+function readAttachedCmd(commandKey: string | undefined): CommandSuggestion | null {
+  if (!commandKey) return null
+  try {
+    const raw = localStorage.getItem(commandKey)
+    if (!raw) return null
+    const parsed: unknown = JSON.parse(raw)
+    if (parsed && typeof parsed === "object" && "command" in parsed) {
+      return parsed as CommandSuggestion
+    }
+  } catch {
+    // corrupt entry — treat as none
+  }
+  return null
 }
 
 /**
@@ -90,6 +109,7 @@ interface Composer {
  */
 function useComposer(
   draftKey: string | undefined,
+  commandKey: string | undefined,
   onSend: ((text: string) => void) | undefined,
 ): Composer {
   // Seed text + caret from the persisted draft ONCE per mount so a remount
@@ -98,7 +118,19 @@ function useComposer(
   const [seed] = useState(() => parseDraft(draftKey))
   const [text, setText] = useState(() => seed.text)
   const [caret, setCaret] = useState(() => seed.selStart)
+  // The attached `/command` (T654) — seeded once from localStorage so it
+  // survives a refresh, prepended to the message on send, at most one.
+  const [attachedCmd, setAttachedCmd] = useState<CommandSuggestion | null>(() =>
+    readAttachedCmd(commandKey),
+  )
   const textareaRef = useRef<HTMLTextAreaElement>(null)
+
+  // Persist the attached command per thread; remove the key when detached.
+  const persistCmd = (c: CommandSuggestion | null) => {
+    if (!commandKey) return
+    if (c) localStorage.setItem(commandKey, JSON.stringify(c))
+    else localStorage.removeItem(commandKey)
+  }
 
   const persistDraft = (t: string, s: number, e: number) => {
     if (!draftKey) return
@@ -134,7 +166,8 @@ function useComposer(
     return line.slice(1)
   }, [text, caret])
 
-  const canSend = (pendingFiles: number) => text.trim().length > 0 || pendingFiles > 0
+  const canSend = (pendingFiles: number) =>
+    text.trim().length > 0 || pendingFiles > 0 || attachedCmd !== null
 
   const applyEdit = (value: string, next: number) => {
     setText(value)
@@ -148,31 +181,53 @@ function useComposer(
     })
   }
 
-  const prefill = (s: CommandSuggestion) => {
-    const base = s.body && s.body.trim().length > 0 ? s.body.trimEnd() : s.command
-    const seeded = `${base}\n`
+  /**
+   * Attach a suggested `/command` as the message prefix (T654). Replaces any
+   * currently-attached command (at most one), persists it durably, and strips
+   * the slash trigger line (bare `/` or partial like `/bo`) out of the draft so
+   * the composer is left clean — the command now rides as a chip, not text.
+   */
+  const attachCmd = (s: CommandSuggestion) => {
     const { start, end } = lineBounds(text, caret)
     const onSlashLine = text.slice(start, end).startsWith("/")
-    const next = onSlashLine ? text.slice(0, start) + seeded + text.slice(end) : seeded
-    const caretPos = onSlashLine ? start + seeded.length : seeded.length
-    setText(next)
-    setCaret(caretPos)
-    persistDraft(next, caretPos, caretPos)
-    requestAnimationFrame(() => {
-      const el = textareaRef.current
-      if (!el) return
-      el.focus()
-      el.setSelectionRange(caretPos, caretPos)
-      autoResize()
-    })
+    if (onSlashLine) {
+      const next = text.slice(0, start) + text.slice(end)
+      setText(next)
+      setCaret(start)
+      persistDraft(next, start, start)
+      requestAnimationFrame(() => {
+        const el = textareaRef.current
+        if (!el) return
+        el.setSelectionRange(start, start)
+        autoResize()
+      })
+    }
+    setAttachedCmd(s)
+    persistCmd(s)
+  }
+
+  /** Remove the attached `/command` (T654). */
+  const detachCmd = () => {
+    setAttachedCmd(null)
+    persistCmd(null)
   }
 
   const handleSubmit = () => {
-    if (text.trim().length === 0 || !onSend) return
-    onSend(text)
+    if (!onSend) return
+    // Prepend the attached command's prompt (T654): its body rides as a prefix,
+    // separated from the user's message by a blank line. Sending is allowed with
+    // only a command attached (empty text). After send, the command detaches.
+    const base = attachedCmd?.body?.trim() ? attachedCmd.body.trimEnd() : ""
+    const msg = base && text.trim() ? `${base}\n\n${text}` : base || text
+    if (msg.trim().length === 0) return
+    onSend(msg)
     setText("")
     setCaret(0)
     persistDraft("", 0, 0)
+    if (attachedCmd) {
+      setAttachedCmd(null)
+      persistCmd(null)
+    }
     requestAnimationFrame(() => {
       const el = textareaRef.current
       if (el) el.style.height = "auto"
@@ -222,137 +277,15 @@ function useComposer(
     caret,
     textareaRef,
     slashPrefix,
+    attachedCmd,
     canSend,
     onChange,
     onSelect,
     handleKeyDown,
     handleSubmit,
-    prefill,
+    attachCmd,
+    detachCmd,
   }
-}
-
-/**
- * The composer's input row — mobile-tuned twin. Same structure as desktop
- * (paperclip + auto-grow textarea + send) with touch-first sizing: 16px textarea
- * font (below 16px iOS Safari auto-zooms the viewport on focus), and 36px tap
- * targets for the paperclip + send.
- */
-function ComposerInputRow({
-  textareaRef,
-  text,
-  sendable,
-  onChange,
-  onSelect,
-  onKeyDown,
-  onSubmit,
-  onAttach,
-}: {
-  textareaRef: React.RefObject<HTMLTextAreaElement | null>
-  text: string
-  sendable: boolean
-  onChange: (e: React.ChangeEvent<HTMLTextAreaElement>) => void
-  onSelect: (e: React.SyntheticEvent<HTMLTextAreaElement>) => void
-  onKeyDown: (e: React.KeyboardEvent<HTMLTextAreaElement>) => void
-  onSubmit: () => void
-  onAttach: ((files: File[]) => void | Promise<void>) | undefined
-}) {
-  const fileInputRef = useRef<HTMLInputElement>(null)
-  // #4 Send-button pop (anime.js): spring the send button in when it becomes
-  // available (iMessage pops it, not a hard cut). Conditionally rendered, so
-  // this fires when `sendable` flips true. Reduced-motion skips it.
-  const sendBtnRef = useRef<HTMLButtonElement>(null)
-  useEffect(() => {
-    const btn = sendBtnRef.current
-    if (!btn || !sendable || prefersReducedMotion()) return
-    animate(btn, {
-      scale: [0, 1],
-      opacity: [0, 1],
-      ease: createSpring({ stiffness: 600, damping: 20 }),
-    })
-  }, [sendable])
-  return (
-    // iMessage-style row: standalone round attach button on the LEFT, then a
-    // single rounded pill holding the textarea with send tucked inside its edge.
-    <div className="flex items-end gap-2">
-      <input
-        ref={fileInputRef}
-        type="file"
-        multiple
-        className="hidden"
-        onChange={(e) => {
-          const files = [...(e.target.files ?? [])]
-          if (files.length > 0) void onAttach?.(files)
-          e.target.value = ""
-        }}
-      />
-      {/* Standalone attach affordance (like iMessage's +), outside the pill. */}
-      <button
-        onClick={() => fileInputRef.current?.click()}
-        disabled={!onAttach}
-        title="Attach files"
-        aria-label="Attach files"
-        className="flex size-9 shrink-0 items-center justify-center rounded-full bg-card/60 text-muted-foreground/70 backdrop-blur-[3px] transition-colors active:bg-muted active:text-(--interactive) disabled:cursor-default disabled:opacity-40"
-      >
-        <Plus className="size-5.5" strokeWidth={2.25} />
-      </button>
-
-      {/* The input pill — thin border, subtle fill, fully rounded. Send button
-          lives INSIDE the pill's right edge (iMessage convention). */}
-      <div className="flex min-w-0 flex-1 items-end gap-1 rounded-[1.35rem] border border-border bg-card/70 py-1 pr-1 pl-3.5 backdrop-blur-[3px] focus-within:border-(--signal)/60">
-        <textarea
-          ref={textareaRef}
-          value={text}
-          onChange={onChange}
-          onSelect={onSelect}
-          onKeyDown={onKeyDown}
-          onPaste={(e) => {
-            const items = [...e.clipboardData.items]
-            const images = items
-              .filter((i) => i.kind === "file" && i.type.startsWith("image/"))
-              .map((i) => i.getAsFile())
-              .filter((f): f is File => f !== null)
-            if (images.length > 0 && onAttach) {
-              e.preventDefault()
-              void onAttach(images)
-            }
-          }}
-          placeholder="Message…"
-          rows={1}
-          className="max-h-[200px] min-h-[30px] min-w-0 flex-1 resize-none self-center bg-transparent py-1 text-[16px] leading-snug text-foreground/90 outline-none placeholder:text-muted-foreground/50"
-        />
-        {/* Send appears only when there's something to send (empty pill stays
-            clean). */}
-        {sendable && (
-          <button
-            ref={sendBtnRef}
-            onClick={onSubmit}
-            aria-label="Send message"
-            className="mb-0.5 flex size-8 shrink-0 items-center justify-center rounded-full bg-(--signal) text-(--primary-foreground) transition-[filter] active:brightness-110"
-          >
-            <ArrowUp className="size-4.5" strokeWidth={2.75} />
-          </button>
-        )}
-      </div>
-    </div>
-  )
-}
-
-/** The turn-status banner element, or null (see {@link resolveComposerBanner}). */
-function ComposerBanner({ banner }: { banner: Banner }) {
-  return (
-    <div
-      className={`mb-2 flex items-center justify-center gap-2 rounded-xl px-3 py-1.5 text-[12px] ${banner.paused ? "bg-amber-500/10 text-amber-600 dark:text-amber-400" : "bg-muted/40 text-muted-foreground"}`}
-    >
-      {banner.paused ? (
-        <Pause className="size-3.5" />
-      ) : banner.working ? (
-        <Loader2 className="size-3.5 animate-spin" style={{ color: banner.color }} />
-      ) : (
-        <Clock className="size-3.5" />
-      )}
-      <span>{banner.text}</span>
-    </div>
-  )
 }
 
 /**
@@ -376,6 +309,7 @@ export function ThreadComposer({
   firstMessage = false,
   onCreateCommand,
   onEditCommand,
+  commandKey,
 }: {
   status: ThreadStatus
   /** true when this is the single thread the agent is currently focused on */
@@ -399,8 +333,11 @@ export function ThreadComposer({
   onEditCommand?: ((s: CommandSuggestion) => void) | undefined
   /** localStorage key for persisting the unsent draft + caret per thread (T304). */
   draftKey?: string | undefined
+  /** localStorage key for the durably-attached `/command` prefix per thread
+   *  (T654). Omit to disable command attachment persistence. */
+  commandKey?: string | undefined
 }) {
-  const composer = useComposer(draftKey, onSend)
+  const composer = useComposer(draftKey, commandKey, onSend)
 
   const userTurn = status === "THEIR_TURN"
   const streaming = status === "ACTIVE"
@@ -424,7 +361,7 @@ export function ThreadComposer({
       const first = filteredSuggestions[0]
       if (first) {
         e.preventDefault()
-        composer.prefill(first)
+        composer.attachCmd(first)
         return
       }
     }
@@ -435,7 +372,7 @@ export function ThreadComposer({
       )
       if (match) {
         e.preventDefault()
-        composer.prefill(match)
+        composer.attachCmd(match)
         return
       }
     }
@@ -447,14 +384,16 @@ export function ThreadComposer({
     <FrostedBottomBar className="px-3 pt-3 pb-[max(1rem,env(safe-area-inset-bottom))]">
       {/* Unified bubble row (T350) — file chips + /command suggestions +
           create-command pill in ONE container. */}
-      {(pendingFiles.length > 0 || commandsActive) && (
+      {(pendingFiles.length > 0 || commandsActive || composer.attachedCmd) && (
         <ComposerBubbles
           files={pendingFiles}
           onRemoveFile={onRemoveFile}
           suggestions={commandsActive ? filteredSuggestions : []}
-          onPick={composer.prefill}
+          onPick={composer.attachCmd}
+          attachedCommand={composer.attachedCmd}
+          onDetachCommand={composer.detachCmd}
           onCreateCommand={commandsActive ? onCreateCommand : undefined}
-          onEditCommand={commandsActive ? onEditCommand : undefined}
+          onEditCommand={onEditCommand}
         />
       )}
       {banner && <ComposerBanner banner={banner} />}
